feat: implementando uma bottom navigation funcional, com as rotas
