+++ pubspec.yaml
@@ -41,6 +41,7 @@ dependencies:
   intl: ^0.17.0
   hooks_riverpod: ^1.0.4
   decimal: ^2.3.0
+  page_transition: ^2.0.9
 
 dev_dependencies:
   flutter_test:
